+++ src/main/java/com/sparta/vicky/config/ControllerUtils.java
@@ -1,4 +1,4 @@
-package com.sparta.vicky.user.controller;
+package com.sparta.vicky.config;
 
 import com.sparta.vicky.user.dto.CommonResponse;
 import lombok.extern.slf4j.Slf4j;
@@ -12,13 +12,13 @@ import java.util.List;
 @Slf4j(topic = "Controller")
 public final class ControllerUtils {
 
-    static ResponseEntity<CommonResponse<?>> getFieldErrorResponseEntity(BindingResult bindingResult, String msg) {
+     public static ResponseEntity<CommonResponse<?>> getFieldErrorResponseEntity(BindingResult bindingResult, String msg) {
         List<FieldError> fieldErrors = bindingResult.getFieldErrors();
         for (FieldError fieldError : fieldErrors) {
             log.error("{} field : {}", fieldError.getField(), fieldError.getDefaultMessage());
         }
 
-        return ResponseEntity.badRequest()
+        return org.springframework.http.ResponseEntity.badRequest()
                 .body(CommonResponse.<List<FieldError>>builder()
                         .statusCode(HttpStatus.BAD_REQUEST.value())
                         .msg(msg)
@@ -26,15 +26,15 @@ public final class ControllerUtils {
                         .build());
     }
 
-    static ResponseEntity<CommonResponse<?>> getBadRequestResponseEntity(Exception e) {
+     public static ResponseEntity<CommonResponse<?>> getBadRequestResponseEntity(Exception e) {
         return ResponseEntity.badRequest().body(CommonResponse.builder()
                 .statusCode(HttpStatus.BAD_REQUEST.value())
                 .msg(e.getMessage())
                 .build());
     }
 
-    static ResponseEntity<CommonResponse<?>> getResponseEntity(Object response, String msg) {
-        return ResponseEntity.ok().body(CommonResponse.builder()
+     public static ResponseEntity<CommonResponse<?>> getResponseEntity(Object response, String msg) {
+        return org.springframework.http.ResponseEntity.ok().body(CommonResponse.builder()
                 .statusCode(HttpStatus.OK.value())
                 .msg(msg)
                 .data(response)
