+++ src/main/java/com/sparta/vicky/user/controller/UserController.java
@@ -14,7 +14,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import static com.sparta.vicky.user.controller.ControllerUtils.*;
+import static com.sparta.vicky.config.ControllerUtils.*;
 
 @RestController
 @RequiredArgsConstructor
